@@ -1,97 +1,93 @@
-name: event_calender_new
-description: "A new Flutter project."
-# The following line prevents the package from being accidentally published to
-# pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+name: we_neighbour
+description: A Flutter app for the We Neighbour community.
+publish_to: 'none'
 
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number is used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-# In Windows, build-name is used as the major, minor, and patch parts
-# of the product and file versions while build-number is used as the build suffix.
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.6.0
+  sdk: ">=3.0.0 <4.0.0"
 
-# Dependencies specify other packages that your package needs in order to work.
-# To automatically upgrade your package dependencies to the latest versions
-# consider running `flutter pub upgrade --major-versions`. Alternatively,
-# dependencies can be manually updated by changing the version numbers below to
-# the latest version available on pub.dev. To see which dependencies have newer
-# versions available, run `flutter pub outdated`.
 dependencies:
+
   flutter:
     sdk: flutter
-  url_launcher: ^6.1.10
+  
+  # UI & Utility Packages
+  provider: ^6.0.2
+  google_fonts: ^6.2.1
+  permission_handler: ^11.0.1
+  flutter_svg: ^2.0.9
+  flutter_carousel_slider: ^1.0.8
   cupertino_icons: ^1.0.8
-  firebase_core: ^2.13.0
-  cloud_firestore: ^4.7.1
+  badges: ^3.1.2
+  image_picker: ^1.1.2
+  shared_preferences: ^2.2.0
+  path_provider: ^2.1.5
+  path: ^1.8.3
+  qr_flutter: ^4.1.0
+  flutter_rating_bar: ^4.0.1
+  url_launcher: ^6.2.2
+  share_plus: ^7.2.1
   intl: ^0.18.1
+  flutter_secure_storage: ^9.0.0
+  add_2_calendar: ^2.2.4
+  
+  # Firebase & Google Services
+  firebase_core: ^3.12.1
+  cloud_firestore: ^5.6.5
+  firebase_auth: ^5.5.1
+  google_sign_in: ^6.2.2
+  googleapis: ^11.4.0
+  googleapis_auth: ^1.4.1
+
+  # Audio & Recording
+  audioplayers: ^6.2.0
+  record: ^5.2.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  http: ^1.3.0  # Ensure this is the latest version
+  flutter_lints: ^3.0.0
+  flutter_launcher_icons: ^0.13.1
 
-  # The "flutter_lints" package below contains a set of recommended lints to
-  # encourage good coding practices. The lint set provided by the package is
-  # activated in the `analysis_options.yaml` file located at the root of your
-  # package. See that file for information about deactivating specific lint
-  # rules and activating additional ones.
+flutter_launcher_icons:
+  android: true
+  ios: true
+  image_path: "assets/images/logo.jpeg"
+  min_sdk_android: 21
+  remove_alpha_ios: true
+  adaptive_icon_background: "#ffffff"
+  adaptive_icon_foreground: "assets/images/logo.jpeg"
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter packages.
 flutter:
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
 
-  # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
-
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/to/resolution-aware-images
-
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/to/asset-from-package
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/to/font-from-package
-
   assets:
-    - assets/calendar_icon.png
-    - assets/amenities_icon.png
-    - assets/health_icon.png
-    - assets/we_neighbour_logo.png
+    - assets/images/white.png
+    - assets/images/facebook.png
+    - assets/images/google.png
+    - assets/images/twitter.png
+    - assets/images/logo.png
+    - assets/icons/amenities.png
+    - assets/icons/visitor.png
+    - assets/icons/calendar.png
+    - assets/icons/maintenance.png
+    - assets/icons/bills.png
+    - assets/icons/chat.png
+    - assets/images/painting.jpg
+    - assets/images/plumbing.jpg
+    - assets/images/carpentry.jpg
+    - assets/images/calendar_icon.png
+    - assets/images/amenities_icon.png
+    - assets/images/health_icon.png
+    - assets/images/we_neighbour_logo.png
+    - assets/images/logo_dark.png
 
+  # Uncomment and update this section if you want to use custom fonts
+  # fonts:
+  #   - family: Roboto
+  #     fonts:
+  #       - asset: assets/fonts/Roboto-Regular.ttf
+  #       - asset: assets/fonts/Roboto-Bold.ttf
+  #         weight: 700
