@@ -1,167 +1,167 @@
-import 'package:flutter/material.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:we_neighbour/screens/chat_screen.dart';
-import '../../constants/colors.dart';
-import 'package:intl/intl.dart';
+// import 'package:flutter/material.dart';
+// import 'package:firebase_auth/firebase_auth.dart';
+// import 'package:cloud_firestore/cloud_firestore.dart';
+// import 'package:we_neighbour/screens/chat_screen.dart';
+// import '../../constants/colors.dart';
+// import 'package:intl/intl.dart';
 
-class ChatListPage extends StatefulWidget {
-  const ChatListPage({Key? key}) : super(key: key);
+// class ChatListPage extends StatefulWidget {
+//   const ChatListPage({Key? key}) : super(key: key);
 
-  @override
-  State<ChatListPage> createState() => _ChatListPageState();
-}
+//   @override
+//   State<ChatListPage> createState() => _ChatListPageState();
+// }
 
-class _ChatListPageState extends State<ChatListPage> {
-  final FirebaseAuth _auth = FirebaseAuth.instance;
-  final TextEditingController _searchController = TextEditingController();
-  String _searchQuery = '';
+// class _ChatListPageState extends State<ChatListPage> {
+//   final FirebaseAuth _auth = FirebaseAuth.instance;
+//   final TextEditingController _searchController = TextEditingController();
+//   String _searchQuery = '';
 
-  @override
-  Widget build(BuildContext context) {
-    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
+//   @override
+//   Widget build(BuildContext context) {
+//     final isDarkMode = Theme.of(context).brightness == Brightness.dark;
 
-    return Scaffold(
-      backgroundColor: isDarkMode ? AppColors.darkBackground : AppColors.background,
-      appBar: AppBar(
-        title: const Text('Chats'),
-        backgroundColor: const Color(0xFF042347),
-        elevation: 0,
-        bottom: PreferredSize(
-          preferredSize: const Size.fromHeight(60),
-          child: Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: TextField(
-              controller: _searchController,
-              decoration: InputDecoration(
-                hintText: 'Search or start a new chat',
-                prefixIcon: const Icon(Icons.search, color: Colors.grey),
-                filled: true,
-                fillColor: isDarkMode ? Colors.grey[800] : Colors.white,
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(30),
-                  borderSide: BorderSide.none,
-                ),
-              ),
-              onChanged: (query) {
-                setState(() {
-                  _searchQuery = query;
-                });
-              },
-            ),
-          ),
-        ),
-      ),
-      body: StreamBuilder<QuerySnapshot>(
-        stream: FirebaseFirestore.instance.collection('users').snapshots(),
-        builder: (context, snapshot) {
-          if (snapshot.hasError) {
-            return Center(child: Text('Error: ${snapshot.error}'));
-          }
+//     return Scaffold(
+//       backgroundColor: isDarkMode ? AppColors.darkBackground : AppColors.background,
+//       appBar: AppBar(
+//         title: const Text('Chats'),
+//         backgroundColor: const Color(0xFF042347),
+//         elevation: 0,
+//         bottom: PreferredSize(
+//           preferredSize: const Size.fromHeight(60),
+//           child: Padding(
+//             padding: const EdgeInsets.all(8.0),
+//             child: TextField(
+//               controller: _searchController,
+//               decoration: InputDecoration(
+//                 hintText: 'Search or start a new chat',
+//                 prefixIcon: const Icon(Icons.search, color: Colors.grey),
+//                 filled: true,
+//                 fillColor: isDarkMode ? Colors.grey[800] : Colors.white,
+//                 border: OutlineInputBorder(
+//                   borderRadius: BorderRadius.circular(30),
+//                   borderSide: BorderSide.none,
+//                 ),
+//               ),
+//               onChanged: (query) {
+//                 setState(() {
+//                   _searchQuery = query;
+//                 });
+//               },
+//             ),
+//           ),
+//         ),
+//       ),
+//       body: StreamBuilder<QuerySnapshot>(
+//         stream: FirebaseFirestore.instance.collection('users').snapshots(),
+//         builder: (context, snapshot) {
+//           if (snapshot.hasError) {
+//             return Center(child: Text('Error: ${snapshot.error}'));
+//           }
 
-          if (snapshot.connectionState == ConnectionState.waiting) {
-            return const Center(child: CircularProgressIndicator());
-          }
+//           if (snapshot.connectionState == ConnectionState.waiting) {
+//             return const Center(child: CircularProgressIndicator());
+//           }
 
-          if (!snapshot.hasData || snapshot.data == null) {
-            return const Center(child: Text('No data available.'));
-          }
+//           if (!snapshot.hasData || snapshot.data == null) {
+//             return const Center(child: Text('No data available.'));
+//           }
 
-          List<DocumentSnapshot> users = snapshot.data!.docs
-              .where((doc) => doc.id != _auth.currentUser!.uid)
-              .toList();
+//           List<DocumentSnapshot> users = snapshot.data!.docs
+//               .where((doc) => doc.id != _auth.currentUser!.uid)
+//               .toList();
 
-          if (_searchQuery.isNotEmpty) {
-            users = users.where((doc) {
-              final userData = doc.data() as Map<String, dynamic>;
-              final userName = userData['email'] ?? '';
-              return userName.toLowerCase().contains(_searchQuery.toLowerCase());
-            }).toList();
-          }
+//           if (_searchQuery.isNotEmpty) {
+//             users = users.where((doc) {
+//               final userData = doc.data() as Map<String, dynamic>;
+//               final userName = userData['email'] ?? '';
+//               return userName.toLowerCase().contains(_searchQuery.toLowerCase());
+//             }).toList();
+//           }
 
-          if (users.isEmpty) {
-            return const Center(child: Text('No users found.'));
-          }
+//           if (users.isEmpty) {
+//             return const Center(child: Text('No users found.'));
+//           }
 
-          return ListView.builder(
-            itemCount: users.length,
-            itemBuilder: (context, index) {
-              final userData = users[index].data() as Map<String, dynamic>;
-              final userId = users[index].id;
-              final userName = userData['email'] ?? 'Unknown User';
-              final userAvatar = userData['avatar'] ?? 'assets/avatars/default_avatar.png';
-              final lastMessage = userData['lastMessage'] ?? ''; // fetch last message
-              final timestamp = userData['timestamp'] != null
-                  ? DateFormat('hh:mm a').format((userData['timestamp'] as Timestamp).toDate())
-                  : ''; // fetch and format timestamp
+//           return ListView.builder(
+//             itemCount: users.length,
+//             itemBuilder: (context, index) {
+//               final userData = users[index].data() as Map<String, dynamic>;
+//               final userId = users[index].id;
+//               final userName = userData['email'] ?? 'Unknown User';
+//               final userAvatar = userData['avatar'] ?? 'assets/avatars/default_avatar.png';
+//               final lastMessage = userData['lastMessage'] ?? ''; // fetch last message
+//               final timestamp = userData['timestamp'] != null
+//                   ? DateFormat('hh:mm a').format((userData['timestamp'] as Timestamp).toDate())
+//                   : ''; // fetch and format timestamp
 
-              return Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 8.0, vertical: 2),
-                child: Card(
-                  color: isDarkMode ? AppColors.darkCardBackground : AppColors.cardBackground,
-                  elevation: 3,
-                  shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
-                  child: ListTile(
-                    leading: CircleAvatar(
-                      radius: 28,
-                      backgroundImage: NetworkImage(userAvatar),
-                    ),
-                    title: Text(
-                      userName,
-                      style: TextStyle(
-                        fontSize: 17,
-                        fontWeight: FontWeight.w600,
-                        color: isDarkMode ? Colors.white : Colors.black,
-                      ),
-                    ),
-                    subtitle: Row(
-                      children: [
-                        Expanded(
-                          child: Text(
-                            lastMessage,
-                            style: TextStyle(
-                              fontSize: 14,
-                              color: isDarkMode ? Colors.grey[400] : Colors.grey[600],
-                            ),
-                            overflow: TextOverflow.ellipsis,
-                          ),
-                        ),
-                      ],
-                    ),
-                    trailing: Column(
-                      mainAxisSize: MainAxisSize.min,
-                      crossAxisAlignment: CrossAxisAlignment.end,
-                      children: [
-                        Text(
-                          timestamp,
-                          style: TextStyle(
-                            fontSize: 12,
-                            color: isDarkMode ? Colors.grey[400] : Colors.grey[600],
-                          ),
-                        ),
-                        const SizedBox(height: 4),
-                        const Icon(Icons.chat_bubble_outline, color: AppColors.primary),
-                      ],
-                    ),
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                          builder: (context) => ChatScreen(
-                            receiverId: userId,
-                            receiverEmail: userName,
-                          ),
-                        ),
-                      );
-                    },
-                    contentPadding: const EdgeInsets.symmetric(vertical: 8, horizontal: 16),
-                  ),
-                ),
-              );
-            },
-          );
-        },
-      ),
-    );
-  }
-}
+//               return Padding(
+//                 padding: const EdgeInsets.symmetric(horizontal: 8.0, vertical: 2),
+//                 child: Card(
+//                   color: isDarkMode ? AppColors.darkCardBackground : AppColors.cardBackground,
+//                   elevation: 3,
+//                   shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+//                   child: ListTile(
+//                     leading: CircleAvatar(
+//                       radius: 28,
+//                       backgroundImage: NetworkImage(userAvatar),
+//                     ),
+//                     title: Text(
+//                       userName,
+//                       style: TextStyle(
+//                         fontSize: 17,
+//                         fontWeight: FontWeight.w600,
+//                         color: isDarkMode ? Colors.white : Colors.black,
+//                       ),
+//                     ),
+//                     subtitle: Row(
+//                       children: [
+//                         Expanded(
+//                           child: Text(
+//                             lastMessage,
+//                             style: TextStyle(
+//                               fontSize: 14,
+//                               color: isDarkMode ? Colors.grey[400] : Colors.grey[600],
+//                             ),
+//                             overflow: TextOverflow.ellipsis,
+//                           ),
+//                         ),
+//                       ],
+//                     ),
+//                     trailing: Column(
+//                       mainAxisSize: MainAxisSize.min,
+//                       crossAxisAlignment: CrossAxisAlignment.end,
+//                       children: [
+//                         Text(
+//                           timestamp,
+//                           style: TextStyle(
+//                             fontSize: 12,
+//                             color: isDarkMode ? Colors.grey[400] : Colors.grey[600],
+//                           ),
+//                         ),
+//                         const SizedBox(height: 4),
+//                         const Icon(Icons.chat_bubble_outline, color: AppColors.primary),
+//                       ],
+//                     ),
+//                     onTap: () {
+//                       Navigator.push(
+//                         context,
+//                         MaterialPageRoute(
+//                           builder: (context) => ChatScreen(
+//                             receiverId: userId,
+//                             receiverEmail: userName,
+//                           ),
+//                         ),
+//                       );
+//                     },
+//                     contentPadding: const EdgeInsets.symmetric(vertical: 8, horizontal: 16),
+//                   ),
+//                 ),
+//               );
+//             },
+//           );
+//         },
+//       ),
+//     );
+//   }
+// }
