@@ -1,33 +1,58 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:google_sign_in/google_sign_in.dart';
+import 'package:flutter/material.dart';
 
 class GoogleSignInService {
-  final GoogleSignIn _googleSignIn = GoogleSignIn();
+  final GoogleSignIn _googleSignIn = GoogleSignIn(
+    scopes: [
+      'email',
+      'https://www.googleapis.com/auth/calendar',
+    ],
+  );
   final FirebaseAuth _auth = FirebaseAuth.instance;
 
+  // Sign in with Google
   Future<UserCredential?> signInWithGoogle() async {
     try {
-      final GoogleSignInAccount? googleSignInAccount = await _googleSignIn.signIn();
-      if (googleSignInAccount != null) {
-        final GoogleSignInAuthentication googleSignInAuthentication =
-            await googleSignInAccount.authentication;
-
-        final OAuthCredential credential = GoogleAuthProvider.credential(
-          accessToken: googleSignInAuthentication.accessToken,
-          idToken: googleSignInAuthentication.idToken,
-        );
-
-        return await _auth.signInWithCredential(credential);
-      }
-      return null;
+      // Trigger the authentication flow
+      final GoogleSignInAccount? googleUser = await _googleSignIn.signIn();
+      
+      // If user cancels the sign-in process
+      if (googleUser == null) return null;
+      
+      // Obtain the auth details from the request
+      final GoogleSignInAuthentication googleAuth = await googleUser.authentication;
+      
+      // Create a new credential
+      final credential = GoogleAuthProvider.credential(
+        accessToken: googleAuth.accessToken,
+        idToken: googleAuth.idToken,
+      );
+      
+      // Sign in to Firebase with the Google credential
+      final userCredential = await _auth.signInWithCredential(credential);
+      
+      return userCredential;
     } catch (e) {
-      print("Error signing in with Google: $e");
+      debugPrint('Error signing in with Google: $e');
       return null;
     }
   }
-
-  Future<void> signOutFromGoogle() async {
+  
+  // Check if user is signed in
+  Future<bool> isSignedIn() async {
+    return await _googleSignIn.isSignedIn();
+  }
+  
+  // Sign out
+  Future<void> signOut() async {
     await _googleSignIn.signOut();
     await _auth.signOut();
   }
+  
+  // Get current user
+  User? getCurrentUser() {
+    return _auth.currentUser;
+  }
 }
+
